Build with a spec's `compile_flags` directive.
