+++ spec/specification/builder_spec.rb
@@ -1,56 +1,24 @@
 require File.expand_path('../../spec_helper', __FILE__)
 
 module Pod
-  describe SpecBuilder do
-    def compare_attributes(first_spec, second_spec, attribute_name)
-      first_spec.attributes_hash[attribute_name].should ==
-        second_spec.attributes_hash[attribute_name]
+  describe Builder do
+    before do
+      @spec = Specification.from_file('spec/fixtures/Builder.podspec')
 
-      %w(ios osx).each do |platform|
-        first_spec.attributes_hash[platform][attribute_name].should ==
-          second_spec.attributes_hash[platform][attribute_name]
-      end
+      @builder = Builder.new(nil, nil, nil, @spec, nil, nil)
     end
 
-    def specification_from_builder(builder)
-      spec_string = builder.spec_metadata
-      spec_string += builder.spec_platform(Platform.ios)
-      spec_string += builder.spec_platform(Platform.osx)
-      spec_string += builder.spec_close
+    describe 'Xcodebuild command' do
 
-      return Specification.from_string(spec_string, 'Builder.podspec')
-    end
-
-  	describe 'Preserve attributes from source specification' do
-      before do
-        @spec = Specification.from_file('spec/fixtures/Builder.podspec')
-        @builder = SpecBuilder.new(@spec, nil, false)
+      it "includes proper compiler flags for iOS" do
+        @builder.expects(:xcodebuild).with('GCC_PREPROCESSOR_DEFINITIONS=\'PodsDummy_Pods_Builder=PodsDummy_PodPackage_Builder\' -DBASE_FLAG -DIOS_FLAG ARCHS="x86_64 i386 arm64 armv7 armv7s"').returns(nil)
+        @builder.compile(Platform.new(:ios))
       end
 
-  		it "preserves platform.frameworks" do
-  			spec = specification_from_builder(@builder)
-        compare_attributes(spec, @spec, 'frameworks')
-  		end
-
-  		it "preserves platform.libraries" do
-  			spec = specification_from_builder(@builder)
-        compare_attributes(spec, @spec, 'libraries')
-  		end
-
-  		it "preserves platform.requires_arc" do
-  			spec = specification_from_builder(@builder)
-        compare_attributes(spec, @spec, 'requires_arc')
-  		end
-
-  		it "preserves platform.deployment_target" do
-  			spec = specification_from_builder(@builder)
-        compare_attributes(spec, @spec, 'deployment_target')
-  		end
-
-  		it "preserves platform.xcconfig" do
-  			spec = specification_from_builder(@builder)
-        compare_attributes(spec, @spec, 'xcconfig')
-  		end
-  	end
+      it "includes proper compiler flags for OSX" do
+        @builder.expects(:xcodebuild).with("GCC_PREPROCESSOR_DEFINITIONS='PodsDummy_Pods_Builder=PodsDummy_PodPackage_Builder' -DBASE_FLAG -DOSX_FLAG").returns(nil)
+        @builder.compile(Platform.new(:osx))
+      end
+    end
   end
 end
